+++ src/client/DeprecatedGUI/Menu_Net_Host.cpp
@@ -1127,7 +1127,7 @@ void Menu_Net_HostLobbyFrame(int mouse)
 						info->addItem(3, "Received: " + itoa(w->getClient()->getChannel()->getIncoming()/1024) + " Kb");
 						info->addItem(4, "Sent: " + itoa(w->getClient()->getChannel()->getOutgoing()/1024) + " Kb");
 						info->addItem(5, "Connected for " + 
-							itoa( (tLX->fCurTime - w->getClient()->getConnectTime()) / 60.0f) + " minutes");
+							itoa( (int)((tLX->fCurTime - w->getClient()->getConnectTime()) / 60.0f)) + " minutes");
 						info->addItem(6, "Total " + 
 							itoa( w->getTotalKills() ) + " kills / " +
 							itoa( w->getTotalDeaths() ) + " deaths / " +
